make-me-a-friggin-pat/requirements.txt
requests>=2.31.0
pycryptodome>=3.18.0
